Fix parsing of template upload commands
Template upload commands have different syntax than upload commands for
non-template packages:

- Only one template can be uploaded at a time.
- Template upload commands do not take git commit hashes, but rather a
  version number.
- Template packages always use the VM package group, never dom0.
- The set of allowed repository names is different.

This needs to be accounted for in the validation code.  Otherwise, it
would be impossible to upload a template.  The shell script validates
the repository name and the version number, but I do not see explicit
checks for the package group or uploading only one template.